Allow users to set the status of their rooms; Parcel out RoomExplorer from TheSideDrawer
